chore: move logic about action conditions to utility module
